Average method sync with Dimo's

+++ Summator/Summator.cs
@@ -14,9 +14,9 @@
             return sum;
         }
 
-        public static int Average(int[] arr)
+        public static double Average(int[] arr)
         {
-            int sum = 0;          
+            double sum = 0;          
 
             for (int i = 0; i < arr.Length; i++)
             {
